refactor: extract implementation of wait condition to module
